% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatmatrix.R
\name{hatmatrix}
\alias{hatmatrix}
\alias{print.hatmatrix}
\title{Derive hat matrix from network meta-analysis}
\usage{
hatmatrix(
  x,
  method = "Ruecker",
  type,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random
)

\method{print}{hatmatrix}(
  x,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random,
  digits = gs("digits"),
  ...
)
}
\arguments{
\item{x}{A \code{\link{netmeta}} object.}

\item{method}{A character string indicating which method is used to
derive the hat matrix. Either \code{"Ruecker"}, \code{"Krahn"} or
\code{"Davies"} (can be abbreviated, see Details).}

\item{type}{A character string indicating which specific hat matrix
should be derived (can be abbreviated, see Details).}

\item{comb.fixed}{A logical indicating whether a hat matrix should
be printed for the fixed effects (common effects) network
meta-analysis.}

\item{comb.random}{A logical indicating whether a hat matrix should
be printed for the random effects network meta-analysis.}

\item{digits}{Minimal number of significant digits, see
\code{print.default}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A list with two hat matrices: \code{fixed} (fixed effect model) and
\code{random} (random effects model).
}
\description{
Auxiliary function to derive hat matrix from network meta-analysis
}
\details{
This auxiliary function can be used to derive various hat matrices
from a network meta-analysis object.

\subsection{Hat matrix by Rücker (2012)}{

This hat matrix is estimated if \code{method = "Ruecker"}.

Let \emph{n} be the number of different treatments (nodes,
vertices) in a network and let \emph{m} be the number of existing
comparisons (edges) between the treatments. If there are only
two-arm studies, \emph{m} is equal to the number of studies,
\emph{k}. Let seTE.adj.fixed and seTE.adj.random be the vectors of
adjusted standard errors under the fixed effect and random effects
model (see \code{\link{netmeta}}). Let \strong{W} be the
\emph{m} x \emph{m} diagonal matrix that contains the inverse
variance 1 / seTE.adj.fixed\eqn{^2} or 1 / seTE.adj.random\eqn{^2}.

The given comparisons define the network structure. Therefrom an
\emph{m} x \emph{n} design matrix X (edge-vertex incidence matrix) is
formed; for more precise information, see Rücker (2012). Moreover,
the \emph{n} x \emph{n} Laplacian matrix \strong{L} and its
Moore-Penrose pseudoinverse \strong{L}\eqn{^+} are calculated (both
matrices play an important role in graph theory and electrical
network theory). Using these matrices, the variances based on both
direct and indirect comparisons can be estimated. The hat matrix
\strong{H} is estimated by \strong{H =
XL}\eqn{^+}\strong{X}\eqn{^T}\strong{W = X(X}\eqn{^T}\strong{W
X)}\eqn{^+}\strong{X}\eqn{^T}\strong{W}.
}

\subsection{Hat matrices by Krahn et al. (2013)}{

One of the following hat matrices is estimated if \code{method
= "Krahn"}.

Use of \code{type = "design"} (default) results in a hat matrix of
dimension \emph{n(n-1)/2} x \emph{d}, where \emph{d} is the sum of
the number of independent comparisons from each design.

Use of \code{type = "studies"} results in a hat matrix of dimension
\emph{n(n-1)/2} x \emph{l}, where \emph{l} is the number of
independent pairwise comparisons, i.e., a three-arm study
contributes two pairwise comparisons.
}

\subsection{Hat matrices by Davies et al. (2021)}{

One of the following hat matrices is estimated if \code{method
= "Davies"}.

Use of \code{type = "short"} (default) results in a hat matrix of
dimension \emph{e} x \emph{e}, where \emph{e} is the number of
(unique) edges (direct comparisons) in the network.

Use of \code{type = "long"} results in a hat matrix of dimension
\emph{n(n-1)/2} x \emph{e}. This hat matrix describes the flow of
evidence through each direct comparison for every possible pair of
treatments (regardless of whether there is direct evidence for this
pair).

Use of \code{type = "full"} results in a hat matrix of dimension
\emph{n(n-1)/2} x \emph{n(n-1)/2}. In comparison to the long hat
matrix, columns of zeroes are added for comparisons that do not
have any direct evidence. This hat matrix is used to calculate the
transition matrices for the random walk in
\code{\link{netcontrib}}.
}
}
\examples{
data(Dong2013)
# Only consider first ten studies for concise output
first10 <- subset(Dong2013, id <= 10)
p1 <- pairwise(treatment, death, randomized, studlab = id,
               data = first10, sm = "OR")
net1 <- netmeta(p1, comb.fixed = FALSE)

hatmatrix(net1)
hatmatrix(net1, method = "k")
hatmatrix(net1, method = "k", type = "studies")
hatmatrix(net1, method = "d")
hatmatrix(net1, method = "d", type = "long")
hatmatrix(net1, method = "d", type = "full")

}
\references{
Davies AL, Papakonstantinou T, Nikolakopoulou A, Rücker G, Galla T
(2021):
Network meta-analysis and random walks.
Available from: http://arxiv.org/abs/2107.02886

Krahn U, Binder H, König J (2013):
A graphical tool for locating inconsistency in network meta-analyses.
\emph{BMC Medical Research Methodology},
\bold{13}, 35

Rücker G (2012):
Network meta-analysis, electrical networks and graph theory.
\emph{Research Synthesis Methods},
\bold{3}, 312--24
}
\seealso{
\code{\link{netmeta}}, \code{\link{netcontrib}},
  \code{\link{netheat}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
