% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.netpath.R
\name{heatplot.netpath}
\alias{heatplot.netpath}
\title{Heat plot for network path}
\usage{
\method{heatplot}{netpath}(x, ...)
}
\arguments{
\item{x}{An object of class \code{netmeta}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
A heat plot where the darker colors indicate a larger conflict between the 
two paths.
}
\description{
Produces a heat plot to visualize the results of a path-based inconsistency
test for a specific comparison, as generated by \code{\link{netpath}}.
The plot shows a standardized matrix derived from the linearly independent
paths connecting the two selected nodes, allowing visual inspection of the
agreement of each path.
}
\examples{
\dontrun{
# Transform data from long arm-based to contrast-based format
#
pw <- pairwise(studlab = study, treat = treatment,
  n = n, mean = mean, sd = sd, data = Senn2013,
  varnames = c("MD", "seMD"))

# Conduct common effects network meta-analysis
#
nma <- netmeta(pw, random = FALSE, nchar.trts = 4)

np <- netpath(nma, node1 = "Placebo", node2 = "Sulfonylurea")
heatplot(np)
}

}
\references{
Tahmasebi NR, Davies AL, Papakonstantinou T, Rücker G,
Nikolakopoulou A (2025):
Path-based approach for detecting and assessing inconsistency in network
meta-analysis: A novel method.
\emph{arXiv}, \doi{https://doi.org/10.48550/arXiv.2506.20364}
}
\seealso{
\code{\link{netpath}}, \code{\link{netmeta}}
}
\author{
Noosheen R. Tahmasebi
  \email{noosheen.rajabzadehtahmasebi@uniklinik-freiburg.de},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{hplot}
