% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.R
\name{heatplot}
\alias{heatplot}
\title{Generic function for heat plots}
\usage{
heatplot(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{\dots}{Additional arguments.}
}
\description{
Generic function for heat plots
}
\details{
For more details, look at the following function to generate heat
plots:
\itemize{
\item \code{\link{heatplot.netmeta}}
\item \code{\link[crossnma]{heatplot.crossnma}} (if installed)
}
}
\examples{
# Only consider first five studies (to reduce runtime of example)
#
studies <- unique(Senn2013$study)
Senn2013.5 <- subset(Senn2013, study \%in\% studies[1:5])

# Transform data from long arm-based to contrast-based format
#
pw <- pairwise(studlab = study, treat = treatment,
  n = n, mean = mean, sd = sd, data = Senn2013.5,
  varnames = c("MD", "seMD"))

# Conduct random effects network meta-analysis with
# placebo as reference treatment
#
nma <- netmeta(pw, common = FALSE, reference = "plac")

# Generate a heat plot (with abbreviated treatment labels)
#
heatplot(nma, nchar.trts = 4) 

}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de
}
}
\keyword{hplot}
