% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpath.R
\name{netpath}
\alias{netpath}
\title{TITLE}
\usage{
netpath(x, random = x$random, node1, node2)
}
\arguments{
\item{x}{A \code{netmeta} object.}

\item{random}{A logical indicating whether the path algorithm is
based on a random effects model.}

\item{node1}{First node.}

\item{node2}{Second node.}
}
\value{
A netpath object.
}
\description{
BRIEF DESCRIPTION (1-2 sentence(s)).
}
\examples{
\dontrun{
data(Senn2013)
nma1 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
  data = Senn2013, sm = "MD", random = FALSE, nchar.trts = 4)

np1 <- netpath(nma1, node1 = "Placebo", node2 = "Sulfonylurea")
np1
}

}
\seealso{
\code{\link{netmeta}}
}
\author{
Noosheen R. Tahmasebi
  \email{noosheen.rajabzadehtahmasebi@uniklinik-freiburg.de},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
