% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.netsplit.R
\name{plot.netsplit}
\alias{plot.netsplit}
\title{Scatter plot to evaluate local inconsistency in network meta-analysis}
\usage{
\method{plot}{netsplit}(
  x,
  type = "Bland-Altman",
  pooled = ifelse(x$x$random, "random", "common"),
  subset = NULL,
  only.reference = x$only.reference,
  level = x$level.ma,
  vertical = TRUE,
  size.points = NULL,
  points.only.inc = FALSE,
  col = "gray50",
  col.inc = "red",
  labels = TRUE,
  labels.only.inc = FALSE,
  size.labels = 5,
  col.labels = col,
  col.labels.inc = col.inc,
  max_overlap = 10,
  nchar.trts = x$nchar.trts,
  text.pooled,
  main,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  tag = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{netsplit}.}

\item{type}{A character string indicating which figure type is to be used.
Either \code{"Bland-Altman"} or \code{"Wilson"}, can be abbreviated.
See Details.}

\item{pooled}{A character string indicating whether results for the
common (\code{"common"}) or random effects model (\code{"random"})
should be plotted. Can be abbreviated.}

\item{subset}{An optional logical vector specifying a subset of
comparisons to consider (must be of same length as the total number of
comparisons).}

\item{only.reference}{A logical indicating whether only comparisons
with the reference group should be considered.}

\item{level}{The level used to calculate confidence intervals for local
local inconsistency.}

\item{vertical}{A logical indicating whether confidence intervals should be
plotted vertically or horizontally; only considered if argument
\code{type = "Wilson"}. The interpretation is the same.}

\item{size.points}{A single numeric defining the point size. If \code{NULL},
point sizes are proportional to the proportion of indirect evidence.}

\item{points.only.inc}{A logical indicating whether (larger) points should
be only shown for comparisons with local inconsistency.}

\item{col}{Colour of points and confidence intervals.}

\item{col.inc}{Colour of points and confidence intervals for comparisons
with local inconsistency.}

\item{labels}{A logical value indicating whether to show comparisons labels,
or an expression or character vector specifying the labels to use.}

\item{labels.only.inc}{A logical indicating whether labels are only shown
for comparisons with local inconsistency.}

\item{size.labels}{A single numeric defining the size of comparison labels.}

\item{col.labels}{Colour of comparison labels.}

\item{col.labels.inc}{Colour of labels for comparisons with local
inconsistency.}

\item{max_overlap}{An integer specifying the maximum number of overlapping
labels allowed in the scatter plot.}

\item{nchar.trts}{A numeric defining the minimum number of
characters used to create unique treatment names in comparisons.}

\item{text.pooled}{A character string used to describe the statistical model.}

\item{main}{Main title.}

\item{xlab}{A label for the x-axis.}

\item{ylab}{A label for the y-axis.}

\item{xlim}{The x limits (min, max) of the plot.}

\item{ylim}{The y limits (min, max) of the plot.}

\item{tag}{The text for the tag label (see \code{\link[ggplot2]{labs}}).}

\item{\dots}{Additional arguments (ignored).}
}
\value{
An object of class \code{"plot.netsplit"}. The object is a list containing
the following components:
\item{plot}{A ggplot2 object.}
\item{data}{Data set used to create the plot.}
}
\description{
Scatter plot to show the difference between direct and indirect evidence in
network meta-analysis to assess local inconsistency.
}
\details{
A scatter plot is drawn in the active graphics window.

By default (argument \code{type = "Bland-Altman"}), a variant of the
Bland-Altman plot is shown (Bland & Altman, 1995). In this case, the
network estimates are shown on the horizontal axis and the difference between
direct and indirect effect estimates on the vertical axis. The confidence
intervals are those from the differences.

If argument \code{type = "Wilson"}, a scatter plot with direct effect
estimates on the horizontal and indirect estimates on the vertical axis
is produced (Wilson et al., 2026). The confidence intervals are
calculated from the indirect (argument \code{vertical = TRUE}) or direct
effect estimates and the standard error of the difference between indirect
and direct effect estimates.

For both plot types, the dashed green line corresponds to comparisons with
equal direct and indirect effect estimates.
}
\examples{
# Examples: example(netsplit)

}
\references{
Bland JM, Altman DG (1995):
Comparing methods of measurement: Why plotting difference against standard
method is misleading.
\emph{The Lancet},
\bold{346}, 1085--87

Wilson H, Schönstein A, Robson S, Bonofiglio F (2026):
A novel approach for visualizing local consistency in
network meta-analysis.
\emph{Research Synthesis Methods},
1--15
}
\seealso{
\code{\link{forest.netsplit}}
}
\author{
Federico Bonofiglio \email{bonostat@gmx.de},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{hplot}
