% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netbind.R
\name{print.netbind}
\alias{print.netbind}
\title{Print method for objects of class netbind}
\usage{
\method{print}{netbind}(
  x,
  common = x$x$common,
  random = x$x$random,
  warn.deprecated = gs("warn.deprecated"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{netbind} or
\code{summary.netbind}.}

\item{common}{A logical indicating whether results for the common
effects model should be printed.}

\item{random}{A logical indicating whether results for the random
effects model should be printed.}

\item{warn.deprecated}{A logical indicating whether warnings should
be printed if deprecated arguments are used.}

\item{\dots}{Additional arguments (to catch deprecated arguments).}
}
\description{
Print method for objects of class \code{netbind}.
}
\examples{
# Examples: example(netbind)

}
\seealso{
\code{\link{netbind}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{print}
