% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.netmeta.R
\name{settings.netmeta}
\alias{settings.netmeta}
\title{Print and change default network meta-analysis settings in R package
\bold{netmeta}}
\usage{
settings.netmeta(..., quietly = TRUE)
}
\arguments{
\item{...}{Arguments to change default settings.}

\item{quietly}{A logical indicating whether information on settings
should be printed.}
}
\description{
Print and change default settings to conduct and print or plot
network meta-analyses in R package \bold{netmeta}.
}
\seealso{
\code{\link[meta]{settings.meta}}, \code{\link[meta]{gs}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
