% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vikor.R
\name{vikor.netposet}
\alias{vikor.netposet}
\alias{vikor.matrix}
\alias{vikor.data.frame}
\alias{vikor}
\alias{print.vikor}
\title{Rank treatments across all outcomes using the VišeKriterijumska Optimizacija
I Kompromisno Rešenje (VIKOR) multi-criteria decision analysis method}
\usage{
\method{vikor}{netposet}(
  x,
  pooled = ifelse(x$random, "random", "common"),
  weights = NULL,
  v = 0.5,
  ...
)

\method{vikor}{matrix}(x, weights = NULL, v = 0.5, method = "SUCRA", ...)

\method{vikor}{data.frame}(x, weights = NULL, v = 0.5, method = "SUCRA", ...)

vikor(x, ...)

\method{print}{vikor}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{\link{netposet}} or a matrix.}

\item{pooled}{A character string indicating whether the VIKOR method should
be based on the common (\code{"common"}) or random effects
model (\code{"random"}). Can be abbreviated.}

\item{weights}{Outcome weights. The weights should always sum to 1. If not
then they are standardized. If NULL, the function will assume equal outcome
weights.}

\item{v}{A scalar from 0 to 1 interpreted as the weight of the decision
making process. Following guidance from the multi-criteria decision
analysis field it is set to 0.5.}

\item{\dots}{Additional arguments (ignored).}

\item{method}{A character string specifying the ranking metric. Either
\code{"P-score"}, \code{"SUCRA"}, \code{"best"}, or
\code{"ranking probabilities"}; can be abbreviated.}

\item{digits}{A numeric specifying the number of digits to print the
ranking matrix Q.}
}
\value{
The function returns a 'vikor' object. This consists of three ranking lists
which are the following:
\itemize{
\item A ranking list Q referring to the ranking when balancing both each
  treatment's overall and worst performance. This is the main ranking list
  of the method. 
\item A ranking list S referring to the ranking in terms of each treatment's
  overall performance.
\item A ranking list R referring to the ranking in terms of penalising each
  treatment's worst performance.
}
In addition to the ranking lists, the function also evaluates the necessary
conditions defined by the VIKOR method and returns a message indicating the
set of compromise solutions.
}
\description{
This function employs the VišeKriterijumska Optimizacija I Kompromisno
Rešenje (VIKOR) method to analyze all outcome-specific ranking lists.
It provides both an amalgamated ranking list and guidance on
which treatments correspond to the best compromise solutions.
}
\details{
This function takes a single mandatory argument, which is either an object
of class \code{\link{netposet}} or a matrix. It then uses the multi-criteria
decision analysis method VišeKriterijumska Optimizacija I Kompromisno
Rešenje (VIKOR) to produce an amalgamated ranking list across
all outcomes (Opricovic & Tzeng, 2004).

The standard VIKOR approach is applied when the \code{method} argument is
equal to \code{"P-score"}, \code{"SUCRA"}, or \code{"best"} in
\code{\link{netposet}}.

The final ranking list is calculated based on treatments
common across all outcomes. Treatments not present across all outcomes are
excluded internally.

Using the argument 'weights' the users can specify the weight that each
outcome should have in the decision making process. For each outcome this
argument should have a value from 0 to 1 while the sum of all outcome
weights should be 1. If the sum of all weights is not 1, then these are
internally standardize to achieve this. The standardized weight values
are returned as a message to the user. Finally, if NULL then equal weights
are assumed across all outcomes.

The argument 'v' specifies the weight of the decision making process.
The VIKOR method is a compromise programming approach that aims to balance
between each treatments overall and worst performance across all outcomes.
The balance between these two criteria is achieved using the parameter 'v'
which takes values from 0 to 1. Values close to 1 will give more weight to
the treatment's overall performance while values close to 0 will give more
weight to penalize the treatment's worst performance. The most common
choice of 'v' is typically 0.5 (default also here), thereby allowing for a
balanced decision making between treatment's overall and worst performance.
}
\examples{
\donttest{
# Define order of treatments in depression data set Linde2015
#
trts <- c("TCA", "SSRI", "SNRI", "NRI",
  "Low-dose SARI", "NaSSa", "rMAO-A", "Hypericum", "Placebo")

# Outcome labels
#
outcomes <- c("Early response", "Early remission")

# (1) Early response
#
pw1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(resp1, resp2, resp3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
#
nma1 <- netmeta(pw1, common = FALSE,
  seq = trts, ref = "Placebo", small.values = "undesirable")

# (2) Early remission
#
pw2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(remi1, remi2, remi3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
#
nma2 <- netmeta(pw2, common = FALSE,
  seq = trts, ref = "Placebo", small.values = "undesirable")

# Partial order of treatment rankings (two outcomes)
#
po12 <- netposet(netrank(nma1), netrank(nma2), outcomes = outcomes)

# Get the best compromise solution across the efficacy outcomes
vikor(po12)

# Use larger weight for response than remission
vikor(po12, weights = c(0.6, 0.4))


# Consider five outcomes
#
# Outcome labels
#
outcomes <- c("Early response", "Early remission",
  "Lost to follow-up", "Lost to follow-up due to AEs",
   "Adverse events (AEs)")

# (3) Loss to follow-up
#
pw3 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss1, loss2, loss3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
#
nma3 <- netmeta(pw3, common = FALSE,
  seq = trts, ref = "Placebo", small.values = "desirable")

# (4) Loss to follow-up due to adverse events
#
pw4 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(loss.ae1, loss.ae2, loss.ae3), n = list(n1, n2, n3),
  studlab = id, data = subset(Linde2015, id != 55), sm = "OR")
#
nma4 <- netmeta(pw4, common = FALSE,
  seq = trts, ref = "Placebo", small.values = "desirable")

# (5) Adverse events
#
pw5 <- pairwise(treat = list(treatment1, treatment2, treatment3),
  event = list(ae1, ae2, ae3), n = list(n1, n2, n3),
  studlab = id, data = Linde2015, sm = "OR")
#
nma5 <- netmeta(pw5, common = FALSE,
  seq = trts, ref = "Placebo", small.values = "desirable")

# Partial order of treatment rankings (based on netrank() objects)
#
po12345 <- netposet(netrank(nma1), netrank(nma2),
  netrank(nma3), netrank(nma4), netrank(nma5), outcomes = outcomes)

# Get the best compromise solution across all outcomes
vikor(po12345)

# Use larger weight for efficacy than safety outcomes
vikor(po12345, weights = c(0.35, 0.35, 0.1, 0.1, 0.1))

# Example using ranking matrix with P-scores
#
# Ribassin-Majed L, Marguet S, Lee A.W., et al. (2017):
# What is the best treatment of locally advanced nasopharyngeal
# carcinoma? An individual patient data network meta-analysis.
# Journal of Clinical Oncology, 35, 498-505
#
# P-scores (from Table 1)
#
pscore.os  <- c(15, 33, 63, 70, 96, 28, 45) / 100
pscore.pfs <- c( 4, 46, 79, 52, 94, 36, 39) / 100
pscore.lc  <- c( 9, 27, 47, 37, 82, 58, 90) / 100
pscore.dc  <- c(16, 76, 95, 48, 72, 32, 10) / 100
#
pscore.matrix <- data.frame(pscore.os, pscore.pfs, pscore.lc, pscore.dc)
rownames(pscore.matrix) <-
  c("RT", "IC-RT", "IC-CRT", "CRT", "CRT-AC", "RT-AC", "IC-RT-AC")
colnames(pscore.matrix) <- c("OS", "PFS", "LC", "DC")
pscore.matrix
#
po <- netposet(pscore.matrix)
vikor(po)
# same result
vikor(pscore.matrix)
}

}
\references{
Opricovic S, Tzeng GH (2004):
Compromise solution by MCDM methods: A comparative analysis of VIKOR and
TOPSIS.
\emph{European Journal of Operational Research},
\bold{156}, 445--55
}
