% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.netpath.R
\name{heatplot.netpath}
\alias{heatplot.netpath}
\title{Heat plot for network path}
\usage{
\method{heatplot}{netpath}(x, ...)
}
\arguments{
\item{x}{An object of class \code{netmeta}.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
Heat plot ...
}
\description{
BRIEF DESCRIPTION (1-2 sentence(s)).
}
\examples{
\dontrun{
data(Senn2013)
nma1 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
  data = Senn2013, sm = "MD", random = FALSE, nchar.trts = 4)
hm <- hatmatrix(nma1, method = "Davies", type = "full")

np1 <- netpath(nma1, hm, node1 = "Placebo", node2 = "Sulfonylurea")
heatplot(np1)
}

}
\author{
Noosheen R. Tahmasebi
  \email{noosheen.rajabzadehtahmasebi@uniklinik-freiburg.de},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{hplot}
