% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linde2016.R
\docType{data}
\name{Linde2016}
\alias{Linde2016}
\title{Network meta-analysis of primary care depression treatments}
\format{A data frame with the following columns:
\tabular{rl}{
\bold{\emph{id}}\tab Study ID \cr
\bold{\emph{lnOR}}\tab Response after treatment (log odds ratio) \cr
\bold{\emph{selnOR}}\tab Standard error of log odds ratio \cr
\bold{\emph{treat1}}\tab First treatment \cr
\bold{\emph{treat2}}\tab Second treatment
}}
\source{
Linde K, Rücker G, Schneider A et al. (2016):
Questionable assumptions hampered interpretation of a network
meta-analysis of primary care depression treatments.
\emph{Journal of Clinical Epidemiology},
\bold{71}, 86--96
}
\description{
Network meta-analysis of 22 treatments (including placebo and usual
care) for the primary care of depression.
}
\examples{
data(Linde2016)

# Only consider studies including Face-to-face PST (to reduce
# runtime of example)
#
face <- subset(Linde2016, id \%in\% c(16, 24, 49, 118))

# Conduct random effects network meta-analysis
#
net1 <- netmeta(lnOR, selnOR, treat1, treat2, id,
                data = face, reference.group = "placebo",
                sm = "OR", comb.fixed = FALSE,
                nchar = 6)
#
summary(net1)

\dontrun{
# Conduct random effects network meta-analysis
#
net2 <- netmeta(lnOR, selnOR, treat1, treat2, id,
                data = Linde2016, reference.group = "placebo",
                sm = "OR", comb.fixed = FALSE,
                nchar = 6)
#
summary(net2)
}
}
\seealso{
\code{\link{netmeta}}, \code{\link{netcomb}}
}
\keyword{datasets}
