% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgraph.R
\name{netgraph}
\alias{netgraph}
\title{Generic function for network graphs}
\usage{
netgraph(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{\dots}{Additional arguments.}
}
\description{
Generic function for network graphs
}
\details{
For more details, look at the following functions to generate
network graphs:
\itemize{
\item \code{\link{netgraph.netmeta}}
\item \code{\link{netgraph.netimpact}}
}
}
\examples{
data(Senn2013)

# Generation of an object of class 'netmeta' with reference
# treatment 'plac'
#
net1 <- netmeta(TE, seTE, treat1, treat2, studlab,
                data = Senn2013, sm = "MD", reference = "plac")

# Network graph with default settings
#
netgraph(net1)

}
\seealso{
\code{\link{netgraph.netmeta}},
  \code{\link{netgraph.netimpact}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de
}
}
\keyword{hplot}
