% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgraph.R
\name{netgraph}
\alias{netgraph}
\title{Generic function for network graphs}
\usage{
netgraph(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{\dots}{Additional arguments.}
}
\description{
Generic function for network graphs
}
\details{
For more details, look at the following functions to generate
network graphs:
\itemize{
\item \code{\link{netgraph.netmeta}}
\item \code{\link{netgraph.netimpact}}
}
}
\seealso{
\code{\link{netgraph.netmeta}},
  \code{\link{netgraph.netimpact}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de
}
}
