% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpairwise.R
\name{netpairwise}
\alias{netpairwise}
\alias{print.netpairwise}
\alias{summary.netpairwise}
\alias{print.summary.netpairwise}
\alias{forest.netpairwise}
\title{Conduct pairwise meta-analyses for all comparisons with direct
evidence}
\usage{
netpairwise(
  x,
  separate = FALSE,
  comb.fixed = x$comb.fixed,
  comb.random = x$comb.random,
  level = x$level,
  level.comb = x$level.comb,
  prediction = x$prediction,
  level.predict = x$level.predict,
  method.tau = x$method.tau,
  ...
)

\method{print}{netpairwise}(x, ...)

\method{summary}{netpairwise}(object, ...)

\method{print}{summary.netpairwise}(x, ...)

\method{forest}{netpairwise}(x, ...)
}
\arguments{
\item{x}{An object of class \code{netmeta} or \code{netpairwise}.}

\item{separate}{A logical indicating whether results for
pairwise comparisons should be printed as separate meta-analyses
or as subgroups which is more concise.}

\item{comb.fixed}{A logical indicating whether a fixed effects
(common effects) network meta-analysis should be conducted.}

\item{comb.random}{A logical indicating whether a random effects
network meta-analysis should be conducted.}

\item{level}{The level used to calculate confidence intervals for
individual comparisons.}

\item{level.comb}{The level used to calculate confidence intervals
for pooled estimates.}

\item{prediction}{A logical indicating whether prediction intervals
should be printed.}

\item{level.predict}{The level used to calculate prediction
intervals for a new study.}

\item{method.tau}{A character string indicating which method is
used to estimate the between-study variance \eqn{\tau^2} and its
square root \eqn{\tau}. Either \code{"DL"}, \code{"REML"}, or
\code{"ML"}, can be abbreviated.}

\item{...}{Additional arguments (passed on to \code{metagen} or
print functions).}

\item{object}{An object of class \code{netpairwise}.}
}
\value{
Either a single \code{\link{metagen}} object with pairwise
comparisons as subgroups or a list with \code{\link{metagen}}
objects for each direct pairwise comparison.
}
\description{
Conduct pairwise meta-analyses for all comparisons with direct
evidence in a network meta-analysis.
}
\details{
Conduct pairwise meta-analyses for all comparisons with direct
evidence in a network meta-analysis. In contrast to
\code{\link{netmeta}} and \code{\link{netsplit}}, unadjusted
standard errors are used in the calculations and the between-study
heterogeneity variance is allowed to differ between comparisons.

The R function \code{\link{metagen}} is called internally.
}
\examples{
data(Senn2013)
oldsets <- settings.meta(digits = 2, digits.tau2 = 2, digits.tau = 2)

# Random effects model
#
net1 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
                data = Senn2013, sm = "MD",
                comb.fixed = FALSE)

# Calculate and print consise results for all pairwise
# meta-analyses
#
np1 <- netpairwise(net1)
np1
print(summary(np1), details.method = FALSE)

\dontrun{
forest(np1)

# Print detailed information for each pairwise comparison
#
np2 <- netpairwise(net1, separate = TRUE)
forest(np2)
}

settings.meta(oldsets)

}
\seealso{
\code{\link{netmeta}}, \code{\link{netsplit}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
