% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpath.R
\name{netpath}
\alias{netpath}
\alias{print.netpath}
\title{Path-based test of inconsistency in network meta-analysis}
\usage{
netpath(x, random = x$random, node1, node2, nchar.trts = x$nchar.trts)

\method{print}{netpath}(
  x,
  nchar.trts = x$nchar.trts,
  sep.trts = x$sep.trts,
  digits.Q = gs("digits.Q"),
  digits.pval.Q = gs("digits.pval.Q"),
  details.methods = gs("details"),
  legend = gs("legend"),
  ...
)
}
\arguments{
\item{x}{A \code{netmeta} object.}

\item{random}{A logical indicating whether the path algorithm is
based on a random effects model.}

\item{node1}{First node.}

\item{node2}{Second node.}

\item{nchar.trts}{A numeric defining the minimum number of
characters used to create unique treatment names (see Details).}

\item{sep.trts}{A character used in comparison names as separator
between treatment labels.}

\item{digits.Q}{Minimal number of significant digits for
heterogeneity statistics, see \code{print.default}.}

\item{digits.pval.Q}{Minimal number of significant digits for
p-value of heterogeneity tests, see \code{print.default}.}

\item{details.methods}{A logical specifying whether details on statistical
methods should be printed.}

\item{legend}{A logical indicating whether a legend should be
printed.}

\item{\dots}{Additional arguments (igored).}
}
\value{
A list of class "netpath" containing the following elements:
\item{results}{Data frame containing results of the Q tests}
\item{A.matrix}{Path-adjacency matrix}
\item{Sigma}{Standardized matrix derived from the linearly independent paths}
\item{theta_p}{Theta p}
\item{random, node1, node2}{As defined above.}
\item{nchar.trts}{As defined above.}
\item{call}{Function call.}
\item{version}{Version of R package netmeta used to create object.}
}
\description{
Performs a path-based test of inconsistency for a specific comparison
(treatment pair) in a network meta-analysis, based on the decomposition of
the hat matrix into independent paths connecting the two nodes.
}
\details{
This function implements a path-based approach to assess inconsistency in a
network meta-analysis (Tahmasebi et al., 2025). Starting from the hat matrix
of the network (as calculated by \code{\link{hatmatrix}} with
\code{method = "Davies"} and \code{type = "full"}), the direct and indirect
evidence contributing to the comparison between \code{node1} and \code{node2}
is decomposed into a set of independent evidence paths. A depth-first search
algorithm is then used to identify these paths. A test statistic Q is
calculated to test whether the estimates derived from these independent paths
are consistent with each other. Under the null hypothesis of consistency, Q
approximately follows a chi-squared distribution with degrees of freedom
equal to the number of independent paths minus one.

Depending on the argument \code{random}, the network estimates are based on
either the common effects model (\code{random = FALSE}) or the random effects
model (\code{random = TRUE}).

In order to get more concise printouts, argument \code{nchar.trts} can be
used to define the minimum number of characters for abbreviated treatment
names (see \code{\link{abbreviate}}, argument \code{minlength}). R function
\code{\link{treats}} is utilised internally to create abbreviated
treatment names.
}
\examples{
\dontrun{
# Transform data from long arm-based to contrast-based format
#
pw <- pairwise(studlab = study, treat = treatment,
  n = n, mean = mean, sd = sd, data = Senn2013,
  varnames = c("MD", "seMD"))

# Conduct common effects network meta-analysis
#
nma <- netmeta(pw, random = FALSE, nchar.trts = 4)

np <- netpath(nma, node1 = "Placebo", node2 = "Sulfonylurea")
np
}

}
\references{
Tahmasebi NR, Davies AL, Papakonstantinou T, Rücker G,
Nikolakopoulou A (2025):
Path-based approach for detecting and assessing inconsistency in network
meta-analysis: A novel method.
\emph{arXiv}, \doi{https://doi.org/10.48550/arXiv.2506.20364}
}
\seealso{
\code{\link{netmeta}}, \code{\link{heatplot.netpath}}
}
\author{
Noosheen R. Tahmasebi
  \email{noosheen.rajabzadehtahmasebi@uniklinik-freiburg.de},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
