% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.netrank.R
\name{plot.netrank}
\alias{plot.netrank}
\title{Plot treatment ranking(s) of network meta-analyses}
\usage{
\method{plot}{netrank}(
  ...,
  name,
  common,
  random,
  seq,
  low = "red",
  mid = "yellow",
  high = "green",
  col = "black",
  main,
  main.size = 14,
  main.col = col,
  main.face = "bold",
  legend = TRUE,
  main.legend = NULL,
  axis.size = 12,
  axis.col = col,
  axis.face = "plain",
  na.value = "grey50",
  angle = 45,
  hjust.x = 1,
  vjust.x = 1,
  hjust.y = 1,
  vjust.y = 0,
  nchar.trts,
  digits = 3,
  digits.mean = 2,
  fixed,
  comb.fixed,
  comb.random,
  warn.deprecated = gs("warn.deprecated")
)
}
\arguments{
\item{...}{One or more \code{\link{netrank}} objects.}

\item{name}{An optional character vector providing descriptive
names for the network meta-analysis objects.}

\item{common}{A logical indicating whether results for the common
effects model should be plotted.}

\item{random}{A logical indicating whether results for the random
effects model should be plotted.}

\item{seq}{A character or numerical vector specifying the sequence
of treatments on the x-axis.}

\item{low}{A character string defining the colour for unfavorable rankings,
see \code{\link[ggplot2]{scale_fill_gradient2}}.}

\item{mid}{A character string defining the colour for average rankings,
see \code{\link[ggplot2]{scale_fill_gradient2}}.}

\item{high}{A character string defining the colour for favorable rankings,
see \code{\link[ggplot2]{scale_fill_gradient2}}.}

\item{col}{Colour of text.}

\item{main}{Title.}

\item{main.size}{Font size of title, see
\code{\link[ggplot2]{element_text}}.}

\item{main.col}{Colour of title, see
\code{\link[ggplot2]{element_text}}.}

\item{main.face}{Font face of title, see
\code{\link[ggplot2]{element_text}}.}

\item{legend}{A logical indicating whether a legend should be
printed.}

\item{main.legend}{Title for legend. If missing, the ranking method is used
for a single ranking method and "Ranking" for different ranking methods
if comparable ranking methods are combined.}

\item{axis.size}{Font size of axis text, see
\code{\link[ggplot2]{element_text}}.}

\item{axis.col}{Colour of axis text, see
\code{\link[ggplot2]{element_text}}.}

\item{axis.face}{Font face of axis text, see
\code{\link[ggplot2]{element_text}}.}

\item{na.value}{Colour for missing values, see
\code{\link[ggplot2]{scale_fill_gradient2}}.}

\item{angle}{Angle for text on x-axis, see
\code{\link[ggplot2]{element_text}}.}

\item{hjust.x}{A numeric between 0 and 1 with horizontal
justification of text on x-axis, see \code{\link[ggplot2]{element_text}}.}

\item{vjust.x}{A numeric between 0 and 1 with vertical
justification of text on x-axis, see
\code{\link[ggplot2]{element_text}}.}

\item{hjust.y}{A numeric between 0 and 1 with horizontal
justification of text on y-axis, see
\code{\link[ggplot2]{element_text}}.}

\item{vjust.y}{A numeric between 0 and 1 with vertical
justification of text on y-axis, see
\code{\link[ggplot2]{element_text}}.}

\item{nchar.trts}{A numeric defining the minimum number of
characters used to create unique treatment names.}

\item{digits}{Minimal number of significant digits for ranking probabilities,
P-scores, and SUCRAs, see \code{\link{print.default}}.}

\item{digits.mean}{Minimal number of significant digits for mean or median
ranks, see \code{\link{print.default}}.}

\item{fixed}{Deprecated argument (replaced by 'common').}

\item{comb.fixed}{Deprecated argument (replaced by 'common').}

\item{comb.random}{Deprecated argument (replaced by 'random').}

\item{warn.deprecated}{A logical indicating whether warnings should
be printed if deprecated arguments are used.}
}
\value{
A ggplot2 object or NULL if no ranking was conducted.
}
\description{
Produce an image plot of treatment ranking(s) generated with R
function \code{netrank}.
}
\details{
This function produces an image plot of network rankings (Palpacuer
et al., 2018, Figure 4). Note, a scatter plot of two network
rankings can be generated with \code{\link{plot.netposet}}.

By default, treatments are ordered by the rankings of the first network
meta-analysis object: decreasing for P-scores, SUCRAs, and probabilities of
being best, and increasing for mean and median ranks. Argument \code{seq}
can be used to specify a different treatment order.

Comparable ranking methods can be shown in the same plot: P-scores, SUCRAs,
and probabilities of being best can be combined, and mean and median ranks
can be combined. These two groups of ranking metrics cannot be combined.
}
\examples{
# Examples: example(netrank)

}
\references{
Palpacuer C, Duprez R, Huneau A, Locher C, Boussageon R, Laviolle
B, et al. (2018):
Pharmacologically controlled drinking in the treatment of alcohol
dependence or alcohol use disorders: a systematic review with
direct and network meta-analyses on nalmefene, naltrexone,
acamprosate, baclofen and topiramate.
\emph{Addiction},
\bold{113}, 220--37
}
\seealso{
\code{\link{netrank}}, \code{\link{netmeta}},
  \code{\link{netposet}}, \code{\link{hasse}},
  \code{\link[metabook]{Linde2015}}
}
\author{
Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}, Clément
  Palpacuer \email{clementpalpacuer@gmail.com}
}
\keyword{hplot}
