% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.netrank.R
\name{plot.netrank}
\alias{plot.netrank}
\title{Plot network ranking(s)}
\usage{
\method{plot}{netrank}(..., name, comb.fixed, comb.random, seq,
  digits = gs("digits.prop"))
}
\arguments{
\item{...}{A single netrank object or a list of netrank objects.}

\item{name}{name}

\item{comb.fixed}{comb.fixed}

\item{comb.random}{comb.random}

\item{seq}{seq}

\item{digits}{digits}
}
\value{
A ggplot2 object:
}
\description{
Plot network ranking(s)
}
\examples{
# Use depression dataset
#
data(Linde2015)

# Define order of treatments
#
trts <- c("TCA", "SSRI", "SNRI", "NRI",
          "Low-dose SARI", "NaSSa", "rMAO-A", "Hypericum",
          "Placebo")

# Outcome labels
#
outcomes <- c("Early response", "Early remission")

# (1) Early response
#
p1 <- pairwise(treat = list(treatment1, treatment2, treatment3),
               event = list(resp1, resp2, resp3),
               n = list(n1, n2, n3),
               studlab = id, data = Linde2015, sm = "OR")
#
net1 <- netmeta(p1, comb.fixed = FALSE,
                seq = trts, ref = "Placebo")

# (2) Early remission
#
p2 <- pairwise(treat = list(treatment1, treatment2, treatment3),
               event = list(remi1, remi2, remi3),
               n = list(n1, n2, n3),
               studlab = id, data = Linde2015, sm = "OR")
#
net2 <- netmeta(p2, comb.fixed = FALSE,
                seq = trts, ref = "Placebo")

# Image plot of treatment rankings (two outcomes)
#
plot(netrank(net1, small.values = "bad"),
     netrank(net2, small.values = "bad"),
     name = outcomes, digits = 2)

\dontrun{
#
# Outcome labels
#
outcomes <- c("Early response", "Early remission",
              "Lost to follow-up", "Lost to follow-up due to AEs",
              "Adverse events (AEs)")

# (3) Loss to follow-up
#
p3 <- pairwise(treat = list(treatment1, treatment2, treatment3),
               event = list(loss1, loss2, loss3),
               n = list(n1, n2, n3),
               studlab = id, data = Linde2015, sm = "OR")
#
net3 <- netmeta(p3, comb.fixed = FALSE,
                seq = trts, ref = "Placebo")

# (4) Loss to follow-up due to adverse events
#
p4 <- pairwise(treat = list(treatment1, treatment2, treatment3),
               event = list(loss.ae1, loss.ae2, loss.ae3),
               n = list(n1, n2, n3),
               studlab = id, data = subset(Linde2015, id != 55),
               sm = "OR")
#
net4 <- netmeta(p4, comb.fixed = FALSE,
                seq = trts, ref = "Placebo")

# (5) Adverse events
#
p5 <- pairwise(treat = list(treatment1, treatment2, treatment3),
               event = list(ae1, ae2, ae3),
               n = list(n1, n2, n3),
               studlab = id, data = Linde2015, sm = "OR")
#
net5 <- netmeta(p5, comb.fixed = FALSE,
                seq = trts, ref = "Placebo")

# Image plot of treatment rankings (two outcomes)
#
plot(netrank(net1, small.values = "bad"),
     netrank(net2, small.values = "bad"),
     netrank(net3, small.values = "good"),
     netrank(net4, small.values = "good"),
     netrank(net5, small.values = "good"),
     name = outcomes, digits = 2)
}

}
\seealso{
\code{\link{netrank}}, \code{\link{netmeta}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
\keyword{hplot}
