% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.netimpact.R
\name{print.netimpact}
\alias{print.netimpact}
\title{Print method for objects of class netimpact}
\usage{
\method{print}{netimpact}(x, ...)
}
\arguments{
\item{x}{An object of class \code{netimpact}.}

\item{\dots}{Additional arguments.}
}
\description{
Print method for objects of class \code{netimpact}.
}
\examples{
data(parkinson)

p1 <- pairwise(list(Treatment1, Treatment2, Treatment3),
               n = list(n1, n2, n3),
               mean = list(y1, y2, y3),
               sd = list(sd1, sd2, sd3),
               data = parkinson, studlab = Study)

net1 <- netmeta(p1)
ni <- netimpact(net1, verbose = TRUE)
ni

}
\seealso{
\code{\link{netimpact}}
}
\author{
Guido Schwarzer \email{sc@imbi.uni-freiburg.de}
}
\keyword{print}
